@@ -1,59 +1,56 @@
 {
-  "name": "muse_matrix",
-  "version": "1.0.0",
+  "name": "new",
+  "version": "0.1.0",
   "private": true,
+  "type": "module",
   "scripts": {
-    "dev": "next dev",
     "build": "next build",
-    "start": "next start",
-    "lint": "next lint"
+    "db:generate": "prisma migrate dev",
+    "db:migrate": "prisma migrate deploy",
+    "db:push": "prisma db push",
+    "db:studio": "prisma studio",
+    "dev": "next dev",
+    "postinstall": "prisma generate",
+    "lint": "next lint",
+    "start": "next start"
   },
   "dependencies": {
-    "@clerk/nextjs": "^5.3.3",
-    "@prisma/client": "^5.7.1",
-    "@radix-ui/react-avatar": "^1.0.4",
-    "@radix-ui/react-dialog": "^1.0.5",
-    "@radix-ui/react-dropdown-menu": "^2.0.6",
-    "@radix-ui/react-slot": "^1.0.2",
-    "@tanstack/react-query": "^5.14.6",
-    "@tanstack/react-query-devtools": "^5.14.7",
-    "@tanstack/react-table": "^8.13.2",
-    "@uploadthing/react": "^6.0.2",
-    "axios": "^1.6.2",
+    "@prisma/client": "^5.14.0",
+    "@radix-ui/react-dialog": "^1.1.1",
+    "@radix-ui/react-dropdown-menu": "^2.1.1",
+    "@radix-ui/react-icons": "^1.3.0",
+    "@radix-ui/react-slot": "^1.1.0",
+    "@t3-oss/env-nextjs": "^0.10.1",
     "class-variance-authority": "^0.7.0",
-    "clsx": "^2.1.0",
-    "framer-motion": "^11.0.3",
-    "lucide-react": "^0.344.0",
-    "next": "^14.1.0",
-    "next-themes": "^0.2.1",
-    "next13-progressbar": "^1.1.1",
-    "nextjs-toploader": "^1.6.4",
-    "react": "^18",
-    "react-colorful": "^5.6.1",
-    "react-dom": "^18",
-    "react-froala-wysiwyg": "^4.1.4",
-    "react-images-uploading": "^3.1.7",
-    "react-paginate": "^8.2.0",
-    "react-quill": "^2.0.0",
-    "react-select": "^5.8.0",
-    "react-spinners": "^0.13.8",
-    "react-toastify": "^10.0.4",
-    "sharp": "^0.33.1",
-    "swiper": "^11.0.5",
-    "tailwind-merge": "^2.2.1",
+    "clsx": "^2.1.1",
+    "geist": "^1.3.0",
+    "lucide-react": "^0.435.0",
+    "next": "^14.2.4",
+    "next-themes": "^0.3.0",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
+    "tailwind-merge": "^2.5.2",
     "tailwindcss-animate": "^1.0.7",
-    "uploadthing": "^6.1.0"
+    "zod": "^3.23.3"
   },
   "devDependencies": {
-    "autoprefixer": "^10.0.1",
-    "encoding": "^0.1.13",
-    "eslint": "^8",
-    "eslint-config-next": "^14.1.0",
-    "postcss": "^8",
-    "prisma": "^5.7.1",
-    "tailwindcss": "^3.3.0"
+    "@types/eslint": "^8.56.10",
+    "@types/node": "^20.14.10",
+    "@types/react": "^18.3.3",
+    "@types/react-dom": "^18.3.0",
+    "@typescript-eslint/eslint-plugin": "^8.1.0",
+    "@typescript-eslint/parser": "^8.1.0",
+    "eslint": "^8.57.0",
+    "eslint-config-next": "^14.2.4",
+    "postcss": "^8.4.39",
+    "prettier": "^3.3.2",
+    "prettier-plugin-tailwindcss": "^0.6.5",
+    "prisma": "^5.14.0",
+    "tailwindcss": "^3.4.3",
+    "typescript": "^5.5.3"
   },
-  "prisma": {
-    "seed": "node prisma/seed.js"
-  }
+  "ct3aMetadata": {
+    "initVersion": "7.37.0"
+  },
+  "packageManager": "npm@10.8.2"
 }
